+++ internal/plugins/resurrection/log.go
@@ -1,7 +1,0 @@
-// Package resurrection ...
-package resurrection
-
-import "github.com/Timwood0x10/ares/internal/logger"
-
-// log is the package-level structured logger.
-var log = logger.Module("resurrection")